Always write matlab data in system byte order
Our data arrays can be in different byte orders. For example, the .acq
data can contain a mix of floats and ints in a byte ordering that
doesn't match the system's ordering. In that case, the float channels
will be in the .acq file's ordering, and the float-ified int channels
will be in the system's native ordering. Chaos ensues.

In this case, making sure our channels are in the system's native byte
order (dtype '=f8') will give us a working .mat file.

+++ bioread/writers/matlabwriter.py
@@ -45,7 +45,7 @@ class MatlabWriter(object):
         for i in range(nc):
             c = data.channels[i]
             chan_dict = {}
-            chan_dict['data'] = c.data
+            chan_dict['data'] = c.data.astype("=f8")
             chan_dict['samples_per_second'] = c.samples_per_second
             chan_dict['name'] = c.name
             chan_dict['frequency_divider'] = c.frequency_divider
